ReOrg 07/01
