+++ .gitignore
@@ -5,3 +5,4 @@ target/
 .settings/
 .DS_STORE
 javadoc
+wikipedia-stats-extractor.iml
