+++ pom.xml
@@ -67,11 +67,6 @@
           <version>1.3.1</version>
       </dependency>
       <dependency>
-          <groupId>org.apache.spark</groupId>
-          <artifactId>spark-hive_2.10</artifactId>
-          <version>1.3.1</version>
-      </dependency>
-      <dependency>
           <groupId>org.dbpedia.extraction</groupId>
           <artifactId>core</artifactId>
           <version>3.9</version> </dependency>
