--- src/main/scala/org/dbpedia/spotlight/util/WikiUtil.scala
@@ -1,121 +1,0 @@
-/**
- *  Copyright 2015 DBpedia Spotlight
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package org.dbpedia.spotlight.util
-
-//import org.dbpedia.extraction.util.RichString.wrapString
-//import org.dbpedia.util.text.uri.UriDecoder
-import java.util.Arrays
-
-/**
- * Contains several utility functions related to WikiText.
- */
-object WikiUtil
-{
-  /**
-   * replace underscores by spaces, replace non-breaking space by normal space, remove
-   * exotic whitespace, normalize duplicate spaces, trim whitespace (any char <= U+0020)
-   * from start and end.
-   *
-   * Also see WikiTitle.parse().
-   *
-   * TODO: better treatment of U+20xx: remove some, replace some by space, others by LF
-   *
-   * FIXME: There is no logic to our decoding / encoding of strings, URIs, etc. It's done
-   * in too many places. We must set a policy and use distinct classes, not generic strings.
-   *
-   * string string possibly using '_' instead of ' '
-   */
-  /*def cleanSpace(string: String): String =
-  {
-    // FIXME: removing these chars may avoid some errors, but also introduces others.
-    // For example, the local name of the 'Project' namespace on fa wikipedia contains U+200C.
-    string.replaceChars("_\u00A0\u200E\u200F\u2028\u202A\u202B\u202C\u3000", "  ").replaceAll(" +", " ").trim
-  }*/
-
-  /**
-   * Replacement string array for StringUtils.escape
-   */
-  //def iriReplacements: Array[String] = StringUtils.replacements('%', "\"#%<>?[\\]^`{|}")
-
-  // store our own private copy of the mutable array
-  //private val replacements = iriReplacements
-
-  /**
-   * Replaces multiple spaces (U+0020) by one, removes spaces from start and end,
-   * replaces spaces by underscores, and percent-encodes the following characters:
-   *
-   *
-   * The result is usable in most parts of a IRI. The ampersand '&' is not escaped though.
-   *
-   * Should only be used for canonical MediaWiki page names. Not for fragments, not for queries.
-   *
-   * TODO: a canonical MediaWiki page name does not contain multiple spaces. We should not
-   * clean spaces but simply throw an exception if the name is not canonical.
-   *
-   * @param name Canonical MediaWiki page name, e.g. 'Émile Zola'
-   */
-  def wikiEncode(name : String) : String=
-  {
-    // TODO: all this replacing is inefficient, one loop over the string would be nicer.
-
-    // replace spaces by underscores.
-    // Note: MediaWiki apparently replaces only spaces by underscores, not other whitespace.
-    var encoded = name.replace(' ', '_');
-
-    // normalize duplicate underscores
-    encoded = encoded.replaceAll("_+", "_");
-
-    // trim underscores from start
-    encoded = encoded.replaceAll("^_", "");
-
-    // trim underscores from end
-    encoded = encoded.replaceAll("_$", "");
-
-    return encoded
-    //StringUtils.escape(encoded, replacements)
-  }
-
-
-  /**
-   * name encoded MediaWiki page name, e.g. '%C3%89mile_Zola'.
-   * Must not include the namespace (e.g. 'Template:').
-   */
-  /*def wikiDecode(name : String) : String =
-  {
-    cleanSpace(UriDecoder.decode(name))
-  }*/
-
-  private val wikiEmphasisRegex1 = "(?s)'''''(.*?)'''''".r
-  private val wikiEmphasisRegex2 = "(?s)'''(.*?)'''".r
-  private val wikiEmphasisRegex3 = "(?s)''(.*?)''".r
-
-  /**
-   * Removes Wiki emphasis.
-   *
-   * @param text
-   * @return The given text without the wiki emphasis
-   */
-  def removeWikiEmphasis(text : String) : String =
-  {
-    // note: I was tempted to replace these three by a single regex,
-    // but it wouldn't really work.
-    var result = wikiEmphasisRegex1.replaceAllIn(text, "$1")
-    result = wikiEmphasisRegex2.replaceAllIn(result, "$1")
-    result = wikiEmphasisRegex3.replaceAllIn(result, "$1")
-    result
-  }
-}
